Refactor create/update rule logic
