+++ apps/web/utils/actions/ai-rule.ts
@@ -2,13 +2,8 @@
 
 import { setUser } from "@sentry/nextjs";
 import { auth } from "@/app/api/auth/[...nextauth]/auth";
-import prisma, { isDuplicateError, isNotFoundError } from "@/utils/prisma";
-import {
-  RuleType,
-  ExecutedRuleStatus,
-  type Action,
-  ActionType,
-} from "@prisma/client";
+import prisma, { isNotFoundError } from "@/utils/prisma";
+import { ExecutedRuleStatus } from "@prisma/client";
 import { getGmailClient } from "@/utils/gmail/client";
 import { aiCreateRule } from "@/utils/ai/rule/create-rule";
 import {
@@ -47,6 +42,7 @@ import { aiFindSnippets } from "@/utils/ai/snippets/find-snippets";
 import { aiRuleFix } from "@/utils/ai/rule/rule-fix";
 import { labelVisibility } from "@/utils/gmail/constants";
 import type { CreateOrUpdateRuleSchemaWithCategories } from "@/utils/ai/rule/create-rule-schema";
+import { safeCreateRule, safeUpdateRule } from "@/utils/rule/rule";
 
 const logger = createScopedLogger("ai-rule");
 
@@ -222,140 +218,6 @@ export const createAutomationAction = withActionInstrumentation<
   if (isActionError(groupIdResult)) return groupIdResult;
   return await safeCreateRule(result, userId, groupIdResult, null);
 });
-
-async function createRule(
-  result: CreateOrUpdateRuleSchemaWithCategories,
-  userId: string,
-  groupId: string | null,
-  categoryIds: string[] | null,
-) {
-  return prisma.rule.create({
-    data: {
-      name: result.name,
-      userId,
-      actions: { createMany: { data: result.actions } },
-      automate: shouldAutomate(result.actions),
-      runOnThreads: shouldRunOnThreads(result.condition),
-      conditionalOperator: result.condition.conditionalOperator,
-      instructions: result.condition.aiInstructions,
-      from: result.condition.static?.from,
-      to: result.condition.static?.to,
-      subject: result.condition.static?.subject,
-      groupId,
-      categoryFilterType: result.condition.categories?.categoryFilterType,
-      categoryFilters: categoryIds
-        ? {
-            connect: categoryIds.map((id) => ({
-              id,
-            })),
-          }
-        : undefined,
-    },
-  });
-}
-
-async function updateRule(
-  ruleId: string,
-  result: CreateOrUpdateRuleSchemaWithCategories,
-  userId: string,
-  groupId: string | null,
-  categoryIds: string[] | null,
-) {
-  return prisma.rule.update({
-    where: { id: ruleId },
-    data: {
-      name: result.name,
-      userId,
-      actions: {
-        deleteMany: {},
-        createMany: { data: result.actions },
-      },
-      automate: shouldAutomate(result.actions),
-      runOnThreads: shouldRunOnThreads(result.condition),
-      conditionalOperator: result.condition.conditionalOperator,
-      instructions: result.condition.aiInstructions,
-      from: result.condition.static?.from,
-      to: result.condition.static?.to,
-      subject: result.condition.static?.subject,
-      groupId,
-      categoryFilterType: result.condition.categories?.categoryFilterType,
-      categoryFilters: categoryIds
-        ? {
-            set: categoryIds.map((id) => ({
-              id,
-            })),
-          }
-        : undefined,
-    },
-  });
-}
-
-export async function safeCreateRule(
-  result: CreateOrUpdateRuleSchemaWithCategories,
-  userId: string,
-  groupId: string | null,
-  categoryIds: string[] | null,
-) {
-  try {
-    const rule = await createRule(result, userId, groupId, categoryIds);
-    return { id: rule.id };
-  } catch (error) {
-    if (isDuplicateError(error, "name")) {
-      // if rule name already exists, create a new rule with a unique name
-      const rule = await createRule(
-        { ...result, name: `${result.name} - ${Date.now()}` },
-        userId,
-        groupId,
-        categoryIds,
-      );
-      return { id: rule.id };
-    }
-
-    logger.error("Error creating rule", {
-      userId,
-      error:
-        error instanceof Error
-          ? { message: error.message, stack: error.stack, name: error.name }
-          : error,
-    });
-
-    return { error: "Error creating rule." };
-  }
-}
-
-export async function safeUpdateRule(
-  ruleId: string,
-  result: CreateOrUpdateRuleSchemaWithCategories,
-  userId: string,
-  groupId: string | null,
-  categoryIds: string[] | null,
-) {
-  try {
-    const rule = await updateRule(ruleId, result, userId, groupId, categoryIds);
-    return { id: rule.id };
-  } catch (error) {
-    if (isDuplicateError(error, "name")) {
-      // if rule name already exists, create a new rule with a unique name
-      const rule = await createRule(
-        { ...result, name: `${result.name} - ${Date.now()}` },
-        userId,
-        groupId,
-        categoryIds,
-      );
-      return { id: rule.id };
-    }
-
-    logger.error("Error updating rule", {
-      userId,
-      error:
-        error instanceof Error
-          ? { message: error.message, stack: error.stack, name: error.name }
-          : error,
-    });
-
-    return { error: "Error updating rule." };
-  }
-}
 
 async function getGroupId(
   result: CreateOrUpdateRuleSchemaWithCategories,
@@ -819,27 +681,6 @@ export const saveRulesPromptAction = withActionInstrumentation(
     };
   },
 );
-
-function shouldAutomate(actions: Pick<Action, "type">[]) {
-  const types = new Set(actions.map((action) => action.type));
-
-  // don't automate replies, forwards, and send emails
-  if (
-    types.has(ActionType.REPLY) ||
-    types.has(ActionType.FORWARD) ||
-    types.has(ActionType.SEND_EMAIL)
-  ) {
-    return false;
-  }
-
-  return true;
-}
-
-// run on threads for static, group, and smart category rules
-// user can enable to run on threads for ai rules themselves
-function shouldRunOnThreads(condition?: { aiInstructions?: string }) {
-  return !condition?.aiInstructions;
-}
 
 /**
  * Generates a rules prompt based on the user's recent email activity and labels.
